# docs/openapi.yaml
openapi: 3.1.0
info:
  title: NexWatch Hub API
  description: >
    Custom API routes exposed by the NexWatch hub (a PocketBase application)
    under /api/custom/*, plus the unauthenticated /healthz liveness probe.
    PocketBase's own built-in routes (record CRUD under /api/collections/*,
    auth, realtime, files, etc.) are not documented here — see the
    PocketBase API reference for those.
  version: "1.0.0"
servers:
  - url: /
security:
  - bearerAuth: []
paths:
...
  /api/custom/notifications/{id}/test:
    post:
      operationId: testNotificationChannel
      summary: Send a test notification to a channel
      description: >
        Dispatches a synthetic "[TEST]" alert through the given
        notification channel to verify its configuration. Requires the
        "operator" role or higher.
      tags: [notifications]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: notification_channels record id.
      responses:
        "200":
          description: Test notification sent.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, enum: [ok] }
                  message: { type: string }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/custom/alerts/{id}/ack:
    post:
      operationId: ackAlert
      summary: Acknowledge a firing alert
      description: >
        Records who acknowledged the alert and when
        (alerts.acknowledged_at/acknowledged_by). An acknowledged alert is
        never re-notified after its cooldown and is never escalated, even if
        it keeps breaching; resolving it later keeps both fields for
        history. Requires the "operator" role or higher.
      tags: [alerts]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: alerts record id.
      responses:
        "200":
          description: Alert acknowledged.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, enum: [ok] }
                  acknowledged_at: { type: string, format: date-time }
                  acknowledged_by: { type: string }
                  acknowledged_alert: { type: string }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/custom/alerts/{id}/unack:
    post:
      operationId: unackAlert
      summary: Clear a previous acknowledgement on an alert
      description: >
        Clears acknowledged_at/acknowledged_by, letting the alert resume
        normal cooldown-gated re-notification and escalation. Requires the
        "operator" role or higher.
      tags: [alerts]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: alerts record id.
      responses:
        "200":
          description: Acknowledgement cleared.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, enum: [ok] }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/custom/silences/active:
    get:
      operationId: listActiveSilences
      summary: List currently-active silences and the agents/checks they cover
      description: >
        Returns every "silences" record whose window currently covers now
        (starts_at <= now < ends_at), each annotated with the ids of every
        agent and check it matches. An agent is matched by agent_id, by an
        overlapping tag, or every agent for a silence with agent_id, tags,
        AND check_ids all empty (global). A check is matched by check_ids,
        by an overlapping tag against the check's own tags, or every check
        for that same global case. List/view is available to any
        authenticated user.
      tags: [silences]
      responses:
        "200":
          description: Active silences.
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: integer }
                  silences:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        name: { type: string }
                        reason: { type: string }
                        agent_id: { type: string }
                        tags: { type: array, items: { type: string } }
                        check_ids: { type: array, items: { type: string } }
                        starts_at: { type: string, format: date-time }
                        ends_at: { type: string, format: date-time }
                        covered_agent_ids: { type: array, items: { type: string } }
                        covered_check_ids: { type: array, items: { type: string } }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        A PocketBase auth token (superuser or "users" record) passed in the
        Authorization header. PocketBase accepts it either as a bare token
        or with a "Bearer " prefix.
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: Authenticated, but the account's role is too low.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: The requested resource does not exist.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    ServerError:
      description: An unexpected server-side error occurred.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
...